requirements.txt
certifi==2017.11.5
cffi==1.11.2
Cython==0.27.3
glfw==1.4.0
imageio==2.2.0
mujoco-py==1.50.1
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
pycparser==2.18
